requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==1.0.1
Pillow==10.1.0
ffmpeg-python==0.2.0
docx2pdf==0.1.8
Werkzeug==3.0.1 